requirements.txt
pynput==1.7.6
requests==2.31.0
python-dotenv==1.0.0 